移除无用解析字段 修复数据类型

+++ WeiboMonitor-netframework/UpdateChecker.cs
@@ -114,7 +114,7 @@ namespace WeiboMonitor
             Enabled = true;
         }
 
-        public void RemoveWeibo(int uid)
+        public void RemoveWeibo(long uid)
         {
             if (TimeLines.Any(x => x.UID == uid) is false)
             {
